+++ test/line_style.cpp
@@ -27,9 +27,9 @@ void test_box::draw() {
   buf[3] = char(sliders[7]->value());
   buf[4] = 0;
   fl_line_style(
-    (int)(choice[0]->item()->user_data()) +
-    (int)(choice[1]->item()->user_data()) +
-    (int)(choice[2]->item()->user_data()),
+    (long)(choice[0]->item()->user_data()) +
+    (long)(choice[1]->item()->user_data()) +
+    (long)(choice[2]->item()->user_data()),
     (int)(sliders[3]->value()),
     buf);
   fl_rect(10,10,w()-20,h()-20);
